simplify result list render, move engine to common dir
https://coveord.atlassian.net/browse/KIT-2600
